+++ src/main/java/maeilmail/subscribequestion/policy/RandomChoicePolicy.java
@@ -1,4 +1,4 @@
-package maeilmail.subscribe.core.policy;
+package maeilmail.subscribequestion.policy;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -6,8 +6,8 @@ import java.util.Random;
 import lombok.RequiredArgsConstructor;
 import maeilmail.question.QuestionQueryService;
 import maeilmail.question.QuestionSummary;
-import maeilmail.subscribe.core.ChoiceQuestionPolicy;
 import maeilmail.subscribe.core.Subscribe;
+import maeilmail.subscribequestion.ChoiceQuestionPolicy;
 import org.springframework.stereotype.Component;
 
 @Component
